feat: 404 page and consolidate widgets into dashboard folder
Adds NotFound page with fallback route. Moves HireHub, UpcomingPanel,
and WidgetGrid from components/widgets into components/dashboard so all
dashboard components live in one place.
