+++ src/pages/Dashboard.jsx
@@ -6,8 +6,8 @@ import ProfileCard from "../components/dashboard/ProfileCard.jsx";
 import ActivityChart from "../components/dashboard/ActivityChart.jsx";
 import JobApplicationBoard from "../components/dashboard/JobApplicationBoard.jsx";
 import ResumeBoard from "../components/dashboard/ResumeBoard.jsx";
-import UpcomingPanel from "../components/widgets/UpcomingPanel.jsx";
-import WidgetGrid from "../components/widgets/WidgetGrid.jsx";
+import UpcomingPanel from "../components/dashboard/UpcomingPanel.jsx";
+import WidgetGrid from "../components/dashboard/WidgetGrid.jsx";
 import { useApplications } from "../hooks/useApplications.js";
 import { useProfile } from "../context/ProfileContext.jsx";
 import "../styles/dashboard.css";
